+++ ricaun.AppBundleTool/AppBundle/AppBundleInfo.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using ricaun.AppBundleTool.PackageContents;
+using System.IO;
 
 namespace ricaun.AppBundleTool.AppBundle
 {
@@ -9,6 +10,7 @@ namespace ricaun.AppBundleTool.AppBundle
         public string PathPackageContents { get; set; }
         public AppBundleFolder AppBundleFolder { get; set; }
         public bool WriteAccess { get; set; }
+        public ApplicationPackage ApplicationPackage { get; set; }
         public AppBundleInfo(string pathBundle)
         {
             PathBundle = pathBundle;
@@ -19,6 +21,8 @@ namespace ricaun.AppBundleTool.AppBundle
                 AppBundleFolder = appBundleFolder;
             }
             WriteAccess = AccessUtils.CheckWriteAccess(pathBundle);
+            ApplicationPackage = ApplicationPackage.Parse(PathPackageContents);
+            ApplicationPackage.Show();
         }
         public bool IsValid()
         {